requirements.txt
fastapi==0.110.1
python-dotenv==1.0.1
typing-extensions==4.11.0
uvicorn[standard]==0.29.0
pytest==8.1.1
# pytest-cov==5.0.0
# pytest-html==4.1.1
# pytest-asyncio==0.23.6
# pytest-trio==0.8.0
httpx==0.27.0
BeautifulReport==0.1.3
# aiocache~=0.12.2
# databases==0.9.0
# asyncpg==0.29.0
funboost==43.2
pydantic-settings==2.2.1
db-libs==1.0
nb-time==1.3
pydantic-yaml==1.3.0
cachetools==5.3.3